+++ lib/static_record/exceptions.rb
@@ -1,6 +1,7 @@
 module StaticRecord
   class RecordNotFound < RuntimeError; end
   class ReservedAttributeName < RuntimeError; end
+  class MissingAttribute < RuntimeError; end
   class NoPrimaryKey < RuntimeError; end
   class UnkownType < RuntimeError; end
   class ClassError < RuntimeError; end