+++ spec/test_app/app/models/articles/article_five.rb
@@ -1,0 +1,5 @@
+class ArticleFive < Article
+  attribute :name, 'Article Five'
+  attribute :important, false
+  attribute :cover, 'article_five.jpg'
+end