+++ spec/test_app/app/models/articles/article_four.rb
@@ -3,4 +3,5 @@ class ArticleFour < Article
   attribute :author, 'Me'
   attribute :rank, 3
   attribute :important, false
+  attribute :cover, Rails.root.join('public', 'articles', 'cover', 'article_four.jpg')
 end