+++ spec/test_app/app/models/articles/article_three.rb
@@ -3,4 +3,5 @@ class ArticleThree < Article
   attribute :author, 'Another author'
   attribute :rank, 1
   attribute :important, true
+  attribute :cover, Rails.root.join('public', 'articles', 'cover', 'article_three.jpg')
 end