+++ spec/test_app/app/models/articles/article_two.rb
@@ -3,4 +3,5 @@ class ArticleTwo < Article
   attribute :author, 'The author'
   attribute :rank, 3
   attribute :important, true
+  attribute :cover, Rails.root.join('public', 'articles', 'cover', 'article_two.jpg')
 end