+++ spec/test_app/app/models/badge.rb
@@ -1,0 +1,6 @@
+class Badge < StaticRecord::Base
+  table     :badges
+  path      Rails.root.join('app', 'models', 'badges', '**', '*.rb')
+  columns   name:         :string,
+            description:  :string
+end