+++ spec/test_app/app/models/badges/badge_one.rb
@@ -1,0 +1,3 @@
+class BadgeOne < Badge
+  attribute :name, 'Badge One'
+end